requirements.txt
requests
python-dotenv
streamlit
pypdf
openai
openai-agents
SpeechRecognition
streamlit-audiorecorder